api/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asposestorage==1.0.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
contourpy==1.3.2
cycler==0.12.1
docopt==0.6.2
fastapi==0.115.2
filelock==3.18.0
fonttools==4.58.0
fsspec==2025.3.2
h11==0.16.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.0
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pipreqs==0.4.13
psutil==7.0.0
pydantic==2.11.4
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
setuptools==80.7.1
six==1.17.0
sniffio==1.3.1
starlette==0.40.0
sympy==1.14.0
tabulate==0.9.0
threadpoolctl==3.6.0
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
yarg==0.1.10
